transition to use zeekrunner
