--- brim/windows/launcher/go.mod
@@ -1,8 +1,0 @@
-module github.com/brimsec/zeek/brim/windows/launcher
-
-go 1.13
-
-require (
-	github.com/alexbrainman/ps v0.0.0-20171229230509-b3e1b4a15894
-	golang.org/x/sys v0.0.0-20200331124033-c3d80250170d // indirect
-)
