+++ brim/windows/zeekrunner/.gitignore
@@ -1,0 +1,1 @@
+zeekrunner.exe
