+++ brim/windows/zeekrunner/go.mod
@@ -1,0 +1,3 @@
+module github.com/brimsec/zeek/brim/windows/zeekrunner
+
+go 1.13
