--- brim/zeek
@@ -1,7 +1,0 @@
-#!/usr/bin/env bash
-
-dir="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null 2>&1 && pwd)"
-
-export ZEEKPATH=$dir/share/zeek:$dir/share/zeek/policy:$dir/share/zeek/site
-export ZEEK_PLUGIN_PATH=$dir/lib/zeek/plugins
-exec "$dir/bin/zeek" "$@"
